+++ lib/dataset/nanmean.cpp
@@ -3,6 +3,7 @@
 /// @file
 /// @author Simon Heybrock
 #include "scipp/dataset/nanmean.h"
+
 #include "scipp/dataset/astype.h"
 #include "scipp/dataset/math.h" // needed by operations_common.h
 #include "scipp/dataset/nansum.h"
@@ -16,7 +17,7 @@ namespace scipp::dataset {
 
 DataArray nanmean(const DataArray &a, const Dim dim) {
   return apply_to_data_and_drop_dim(
-      a, [](auto &&... _) { return nanmean(_...); }, dim, a.masks());
+      a, [](auto &&..._) { return nanmean(_...); }, dim, a.masks());
 }
 
 DataArray nanmean(const DataArray &a) {
@@ -25,11 +26,10 @@ DataArray nanmean(const DataArray &a) {
 
 Dataset nanmean(const Dataset &d, const Dim dim) {
   return apply_to_items(
-      d, [](auto &&... _) { return nanmean(_...); }, dim);
+      d, [](auto &&..._) { return nanmean(_...); }, dim);
 }
 
 Dataset nanmean(const Dataset &d) {
-  return apply_to_items(d, [](auto &&... _) { return nanmean(_...); });
+  return apply_to_items(d, [](auto &&..._) { return nanmean(_...); });
 }
-
 } // namespace scipp::dataset