+++ lib/dataset/test/bin_test.cpp
@@ -11,7 +11,6 @@
 #include "scipp/dataset/string.h"
 #include "scipp/variable/arithmetic.h"
 #include "scipp/variable/comparison.h"
-#include "scipp/variable/misc_operations.h"
 #include "scipp/variable/reduction.h"
 #include "scipp/variable/util.h"
 
@@ -21,7 +20,7 @@ using testdata::make_table;
 
 template <class Coords> class DataArrayBinTest : public ::testing::Test {
 protected:
-  template <size_t I, class... Vals> auto make_coord(Dim dim, Vals &&... vals) {
+  template <size_t I, class... Vals> auto make_coord(Dim dim, Vals &&...vals) {
     using Coord = std::tuple_element_t<I, Coords>;
     return makeVariable<Coord>(Dims{dim}, Shape{sizeof...(vals)},
                                Values{static_cast<Coord>(vals)...});